requirements.txt
asgiref==3.2.10
attrs==19.3.0
django-heroku==0.3.1
autobahn==20.6.2
Automat==20.2.0
cffi==1.14.0
channels==2.4.0
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
Django==3.0.8
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
hyperlink==19.0.0
idna==2.10
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==19.1.0
pytz==2020.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
zope.interface==5.1.0